fixed and allowed save file overwrite

+++ Assets/Scripts/LevelEditor/SaveLoad_Handler_Script.cs
@@ -23,9 +23,24 @@ public class SaveLoad_Handler_Script : MonoBehaviour
          * Tile 2: transfer data
          * 
          */
-        StreamWriter sr = new StreamWriter(contentBar.MapFilePath + saveAsName.text + ".MapData", false);
-        Debug.Log(contentBar.MapFilePath + saveAsName.text + ".MapData");
-        sr.WriteLine(worldHandler.getBuildLayers().width + "," + worldHandler.getBuildLayers().height + "," + mapName.text);
+        if ((worldHandler.getBuildLayers() == null || saveAsName.text == null || saveAsName.text.Length == 0) && existingFileName.value == 0)
+        {
+            return;
+        }
+        StreamWriter sr;
+        if (existingFileName.value == 0)
+        {
+            sr = new StreamWriter(contentBar.MapFilePath + saveAsName.text + ".MapData", false);
+            Debug.Log(contentBar.MapFilePath + saveAsName.text + ".MapData");
+        }
+        else
+        {
+            sr = new StreamWriter(contentBar.MapFilePath + existingFileName.options[existingFileName.value].text + ".MapData", false);
+            Debug.Log(contentBar.MapFilePath + existingFileName.options[existingFileName.value].text + ".MapData");
+        }
+        sr.WriteLine(worldHandler.getBuildLayers().width + "," + 
+            worldHandler.getBuildLayers().height + "," + 
+            mapName.text);
         for (int i = 0; i < worldHandler.getBuildLayers().width; i++ )
         {
             for (int j = 0; j < worldHandler.getBuildLayers().height; j++)
